% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreRes.R
\name{CoreRes}
\alias{CoreRes}
\title{This function calulates the results after nucleus simulation}
\usage{
CoreRes(var_name = var_name)
}
\arguments{
\item{var_name}{name of variable combination}
}
\description{
This function calulates the results after nucleus simulation
}
