% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreSim.R
\name{CoreSim}
\alias{CoreSim}
\title{This function simulaitng nucleus breeding process}
\usage{
CoreSim(nvar)
}
\arguments{
\item{nvar}{the num of variable combination}
}
\description{
This function simulaitng nucleus breeding process
}
