% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmatrix.R
\name{g_matrix_cal}
\alias{g_matrix_cal}
\title{construct G matrix}
\usage{
g_matrix_cal(snp_file = "Geno", inverse = FALSE, prefix_name = NULL)
}
\arguments{
\item{snp_file}{which plink file need to calculate g matrix, .raw format}

\item{inverse}{need to calculate inverse g matrix or not}

\item{prefix_name}{the prefix name of output file's path}
}
\description{
construct G matrix
}
