% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptRes.R
\name{getOptRes}
\alias{getOptRes}
\title{This function extracts opt outlog results}
\usage{
getOptRes(res = res, out_path = out_path)
}
\arguments{
\item{res}{the output results of runOpt()}

\item{out_path}{path of output results}
}
\value{
a csv file of bayesion optimization's results
}
\description{
This function extracts opt outlog results
}
