% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HM.R
\name{HM}
\alias{HM}
\title{This function makes homozygosity/heterozygosity mating plan}
\usage{
HM(Mate = Mate, path = "plans.txt", maxF = maxF, Sor = Sor)
}
\arguments{
\item{Mate}{the mating method, HOMO:homozygosity, HE:heterozygosity}

\item{path}{the path of mating plan}

\item{maxF}{the allowed maximum inbreeding coeffecients between parents. Default is 999.}

\item{Sor}{the number of dams mate with one sire}
}
\value{
a list of crossplan matrix and inbreeding values
}
\description{
This function makes homozygosity/heterozygosity mating plan
}
