% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mate.R
\name{Mate_file}
\alias{Mate_file}
\title{This function prepares files for performing further mating}
\usage{
Mate_file()
}
\arguments{
\item{NULL}{}
}
\description{
This function prepares files for performing further mating
}
