% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mate_rand.R
\name{Mate_rand}
\alias{Mate_rand}
\title{This function makes random mate}
\usage{
Mate_rand(maxF = maxF)
}
\arguments{
\item{maxF}{the allowed maximum inbreeding coeffecients between parents. Default is 999.}
}
\value{
a list of crossplan and inbreeding values
}
\description{
This function makes random mate
}
