% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptSim.R
\name{OptSim}
\alias{OptSim}
\title{This function simulates breeding process for optimization}
\usage{
OptSim(x)
}
\arguments{
\item{x}{a vector contains the value of Asec,Bsec,Csec,YABd,NAd,NBd,NCd}
}
\value{
the breeding sysytem's profit of variable combination
}
\description{
This function simulates breeding process for optimization
}
