% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCore.R
\name{runCore}
\alias{runCore}
\title{This function runs core breeding programs task}
\usage{
runCore(prm_path = prm_path)
}
\arguments{
\item{prm_path}{the full path to the prm file}
}
\description{
This function runs core breeding programs task
}
