% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MC.R
\name{runMC}
\alias{runMC}
\title{generate mate plans using MC method}
\usage{
runMC(Nrows = Nrows, path = getwd())
}
\arguments{
\item{Nrows}{the number of plans file's rows}

\item{path}{the path of plans file}
}
\description{
generate mate plans using MC method
}
