% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOpt.R
\name{runOpt}
\alias{runOpt}
\title{This function runs breeding programs' optimization task}
\usage{
runOpt(prm_path = prm_path)
}
\arguments{
\item{prm_path}{the full path to the prm file}
}
\description{
This function runs breeding programs' optimization task
}
