% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWhole.R
\name{runWhole}
\alias{runWhole}
\title{This function simlates whole system's breeding programs task}
\usage{
runWhole(prm_path = prm_path)
}
\arguments{
\item{prm_path}{the full path to the prm file}
}
\description{
This function simlates whole system's breeding programs task
}
