% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAcc.R
\name{SimAcc}
\alias{SimAcc}
\title{This function simulates ebv according accuracy}
\usage{
SimAcc()
}
\arguments{
\item{NULL}{}
}
\value{
a dataframe of simulated ebv
}
\description{
This function simulates ebv according accuracy
}
