% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraceGeno.R
\name{TraceGeno}
\alias{TraceGeno}
\title{This function traces Geno for n generation}
\usage{
TraceGeno(ped = Ped, Geno = Geno, n = 3, id = famselect@id)
}
\arguments{
\item{ped}{pedigree file}

\item{Geno}{genotype file}

\item{n}{the number of generation want to trace}

\item{id}{the vector of individuals that wanted to trace generations}
}
\value{
return a genotype file contain exact generations
}
\description{
This function traces Geno for n generation
}
