% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WholeRes.R
\name{WholeRes}
\alias{WholeRes}
\title{This function gets whole breeding task's results}
\usage{
WholeRes(var_name = var_name)
}
\arguments{
\item{var_name}{the var_name of variable combination}
}
\description{
This function gets whole breeding task's results
}
