% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalWeight.R
\name{CalWeight}
\alias{CalWeight}
\title{calculating traits' weight for aggregate index}
\usage{
CalWeight(pz)
}
\arguments{
\item{pz}{the name of the breed}
}
\value{
return the weight of each trait
}
\description{
calculating traits' weight for aggregate index
}
