% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/founder.R
\name{Founders}
\alias{Founders}
\title{create founder populations for breeds}
\usage{
Founders(Breed = Breed)
}
\arguments{
\item{Breed}{a vector of breeds' name}
}
\description{
create founder populations for breeds
}
