% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_dir.R
\name{FUN_dir}
\alias{FUN_dir}
\title{generate dir file for running dmu}
\usage{
FUN_dir(phe, path, Method = Method)
}
\arguments{
\item{phe}{the phenotype name}

\item{path}{the out path of the dir file}

\item{Method}{the method of estimating breeding values ('blup' or 'ssgblup')}
}
\description{
generate dir file for running dmu
}
