% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_gap.R
\name{fun_gap}
\alias{fun_gap}
\title{calculate gene flow when only A has secondary group}
\usage{
fun_gap(par1 = Ctype, par2 = Time)
}
\arguments{
\item{par1}{cross type of whole system: 11,12,13,21,22,23,24,31,32,33}

\item{par2}{Time of whole selection years}
}
\description{
calculate gene flow when only A has secondary group
}
