% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptRes.R
\name{getOptRes}
\alias{getOptRes}
\title{extract opt outlog results}
\usage{
getOptRes(log_name = log_name, out_path = out_path)
}
\arguments{
\item{log_name}{whole path of opt out log}

\item{out_path}{path of output results}
}
\description{
extract opt outlog results
}
