% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HM.R
\name{HM}
\alias{HM}
\title{run HM mate}
\usage{
HM(Mate = Mate, path = "plans.txt", maxF = maxF, Sor = Sor)
}
\arguments{
\item{NULL}{}
}
\value{
a list of crossplan and inbreeding values
}
\description{
run HM mate
}
