% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptSim.R
\name{OptSim}
\alias{OptSim}
\title{simulation of breeding process for optimization}
\usage{
OptSim(x)
}
\arguments{
\item{Asec}{service time of A's secondary group}

\item{Bsec}{service time of B's secondary group}

\item{Csec}{service time of C's secondary group}

\item{YABd}{service time of hybrids AB}

\item{NAd}{service number of nucleus A's female stocks}

\item{NBd}{service number of nucleus B's female stocks}

\item{NCd}{service number of nucleus C's female stocks}
}
\value{
the breeding sysytem's profit of variable combination
}
\description{
simulation of breeding process for optimization
}
