% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrm4.R
\name{readPrm}
\alias{readPrm}
\title{read parameter file}
\usage{
readPrm(prm_path = prm_path, task = task)
}
\arguments{
\item{prm_path}{path of parameter file}

\item{task}{number of analyse (1,2,3)}
}
\description{
read parameter file
}
