% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TracePed.R
\name{TracePed}
\alias{TracePed}
\title{Tracing ped for n generation}
\usage{
TracePed(ped = Ped, n = 3, id = famselect@id)
}
\arguments{
\item{ped}{pedigree file}

\item{n}{the number of generation want to trace}

\item{id}{the vector of individuals that wanted to trace pedigree}
}
\value{
return a pedigree file contain exact generations
}
\description{
Tracing ped for n generation
}
