% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WholeSim.R
\name{WholeSim}
\alias{WholeSim}
\title{run whole breeding system's simulation}
\usage{
WholeSim(nvar)
}
\arguments{
\item{nvar}{the num of variable combination}
}
\description{
run whole breeding system's simulation
}
